Final adjusts

+++ src/main/java/com/fiap/project/EcomerceProject/dto/ClientsDTO.java
@@ -13,7 +13,7 @@ public class ClientsDTO implements Serializable{
 	private static final long serialVersionUID = 1L;
 
 	
-	private Integer id;
+	private Long id;
 
 	private String name;
 
@@ -68,17 +68,17 @@ public class ClientsDTO implements Serializable{
 
 
 
-	public Integer getId() {
+	
+
+
+
+	public Long getId() {
 		return id;
 	}
 
-
-
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
-
-
 
 	public String getName() {
 		return name;
