+++ src/main/java/com/fiap/project/EcomerceProject/entities/Clients.java
@@ -23,7 +23,7 @@ public class Clients implements Serializable{
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Integer id;
+	private Long id;
 
 	private String name;
 
@@ -58,12 +58,14 @@ public class Clients implements Serializable{
 		this.postalCode = postalCode;
 		this.city = city;
 	}
+  
+    
 
-	public Integer getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
